更新

+++ src/main/java/com/dx/pojo/dto/CoinDTO.java
@@ -3,6 +3,7 @@ package com.dx.pojo.dto;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 public class CoinDTO implements Serializable {
@@ -22,7 +23,7 @@ public class CoinDTO implements Serializable {
     /**
      * 最小收款数
      */
-    private Integer minNum;
+    private BigDecimal minNum;
     /**
      * 主网名称
      */
